@@ -18,32 +18,21 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  auto_size_text:
-  built_collection: ^5.0.0
-  built_value: ^8.0.6
   cached_network_image: ^3.1.0
-  cloud_firestore: ^2.4.0
-  firebase_auth: ^3.0.1
-  firebase_core: ^1.4.0
-  flutter_datetime_picker: 1.5.0
-  flutter_spinkit: ^4.1.2+1
+  flutter_datetime_picker: ^1.5.1
+  flutter_spinkit: ^5.0.0
   font_awesome_flutter:
   geolocator:
   google_fonts: ^2.1.0
-  google_sign_in: ^4.5.9
+  google_sign_in: ^5.0.7
   intl:
-  json_serializable: ^4.0.1
   page_transition:
-  rxdart: ^0.26.0
   sign_in_with_apple: ^3.0.0
-  timeago:
-  url_launcher:
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -51,11 +40,12 @@ dependencies:
   duration: ^3.0.6
   flutter_iconpicker: ^3.0.2
   flutter_svg: ^0.22.0
+  backendless_sdk: ^7.0.0
+  timeago: ^3.1.0
+  auto_size_text: 3.0.0-nullsafety.0
+  url_launcher: ^6.0.9
 
 dev_dependencies:
-  build_runner: ^1.11.5
-  built_value_generator: ^8.0.4
-
   flutter_test:
     sdk: flutter
 
