@@ -1,6 +1,6 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:flutter/material.dart';
-import 'package:auto_size_text/auto_size_text.dart';
 
 class FFButtonOptions {
   const FFButtonOptions({
@@ -20,35 +20,35 @@ class FFButtonOptions {
     this.borderSide,
   });
 
-  final TextStyle textStyle;
-  final double elevation;
-  final double height;
-  final double width;
-  final EdgeInsetsGeometry padding;
-  final Color color;
-  final Color disabledColor;
-  final Color disabledTextColor;
-  final Color splashColor;
-  final double iconSize;
-  final Color iconColor;
-  final EdgeInsetsGeometry iconPadding;
-  final double borderRadius;
-  final BorderSide borderSide;
+  final TextStyle? textStyle;
+  final double? elevation;
+  final double? height;
+  final double? width;
+  final EdgeInsetsGeometry? padding;
+  final Color? color;
+  final Color? disabledColor;
+  final Color? disabledTextColor;
+  final Color? splashColor;
+  final double? iconSize;
+  final Color? iconColor;
+  final EdgeInsetsGeometry? iconPadding;
+  final double? borderRadius;
+  final BorderSide? borderSide;
 }
 
 class FFButtonWidget extends StatelessWidget {
   const FFButtonWidget({
-    Key key,
-    @required this.text,
-    @required this.onPressed,
+    Key? key,
+    required this.text,
+    required this.onPressed,
     this.icon,
     this.iconData,
-    @required this.options,
+    required this.options,
   }) : super(key: key);
 
   final String text;
-  final Widget icon;
-  final IconData iconData;
+  final Widget? icon;
+  final IconData? iconData;
   final VoidCallback onPressed;
   final FFButtonOptions options;
 
@@ -65,29 +65,30 @@ class FFButtonWidget extends StatelessWidget {
       return Container(
         height: options.height,
         width: options.width,
-        child: RaisedButton.icon(
+        child: ElevatedButton.icon(
           icon: Padding(
             padding: options.iconPadding ?? EdgeInsets.zero,
             child: icon ??
                 FaIcon(
                   iconData,
                   size: options.iconSize,
-                  color: options.iconColor ?? options.textStyle.color,
+                  color: options.iconColor ?? options.textStyle?.color,
                 ),
           ),
           label: textWidget,
           onPressed: onPressed,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(options.borderRadius),
-            side: options.borderSide ?? BorderSide.none,
-          ),
-          color: options.color,
-          colorBrightness: ThemeData.estimateBrightnessForColor(options.color),
-          textColor: options.textStyle.color,
-          disabledColor: options.disabledColor,
-          disabledTextColor: options.disabledTextColor,
-          elevation: options.elevation,
-          splashColor: options.splashColor,
+          style: ElevatedButton.styleFrom(
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(options.borderRadius!),
+                side: options.borderSide ?? BorderSide.none,
+              ),
+              elevation: options.elevation,
+              textStyle: TextStyle(
+                color: options.textStyle?.color,
+              ),
+              onPrimary: options.textStyle?.color,
+              primary: options.textStyle?.backgroundColor,
+              onSurface: options.disabledColor),
         ),
       );
     }
@@ -95,19 +96,22 @@ class FFButtonWidget extends StatelessWidget {
     return Container(
       height: options.height,
       width: options.width,
-      child: RaisedButton(
+      child: ElevatedButton(
         onPressed: onPressed,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(options.borderRadius ?? 28),
-          side: options.borderSide ?? BorderSide.none,
+        style: ElevatedButton.styleFrom(
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(options.borderRadius ?? 28),
+            side: options.borderSide ?? BorderSide.none,
+          ),
+          textStyle: TextStyle(
+            color: options.textStyle?.color,
+          ),
+          primary: options.textStyle?.backgroundColor,
+          onPrimary: options.textStyle?.color,
+          onSurface: options.disabledColor,
+          padding: options.padding,
+          elevation: options.elevation,
         ),
-        textColor: options.textStyle.color,
-        color: options.color,
-        colorBrightness: ThemeData.estimateBrightnessForColor(options.color),
-        disabledColor: options.disabledColor,
-        disabledTextColor: options.disabledTextColor,
-        padding: options.padding,
-        elevation: options.elevation,
         child: textWidget,
       ),
     );
