@@ -1,14 +1,9 @@
-import '../components/add_habit_form_widget.dart';
-import '../flutter_flow/flutter_flow_theme.dart';
+import 'package:habify/components/widgets/add_habit_form_widget.dart';
+import 'package:habify/flutter_flow/flutter_flow_theme.dart';
 import 'package:flutter/material.dart';
 
 class AddHabitWidget extends StatefulWidget {
-  AddHabitWidget({
-    Key key,
-    this.isGoodStyle,
-  }) : super(key: key);
-
-  final bool isGoodStyle;
+  AddHabitWidget({Key? key}) : super(key: key);
 
   @override
   _AddHabitWidgetState createState() => _AddHabitWidgetState();
