@@ -1,80 +1,9 @@
-import 'package:built_value/serializer.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+import 'package:habify/entities/repository.dart';
+import 'package:habify/entities/secret.dart';
 
-import '../flutter_flow/flutter_flow_util.dart';
-
-import 'schema/habits_record.dart';
-import 'schema/users_record.dart';
-import 'schema/types_record.dart';
-import 'schema/serializers.dart';
-
-export 'package:cloud_firestore/cloud_firestore.dart';
-export 'schema/index.dart';
-export 'schema/serializers.dart';
-
-export 'schema/habits_record.dart';
-export 'schema/users_record.dart';
-export 'schema/types_record.dart';
-
-Stream<List<HabitsRecord>> queryHabitsRecord(
-        {Query Function(Query) queryBuilder,
-        int limit = -1,
-        bool singleRecord = false}) =>
-    queryCollection(HabitsRecord.collection, HabitsRecord.serializer,
-        queryBuilder: queryBuilder, limit: limit, singleRecord: singleRecord);
-
-Stream<List<UsersRecord>> queryUsersRecord(
-        {Query Function(Query) queryBuilder,
-        int limit = -1,
-        bool singleRecord = false}) =>
-    queryCollection(UsersRecord.collection, UsersRecord.serializer,
-        queryBuilder: queryBuilder, limit: limit, singleRecord: singleRecord);
-
-Stream<List<TypesRecord>> queryTypesRecord(
-        {Query Function(Query) queryBuilder,
-        int limit = -1,
-        bool singleRecord = false}) =>
-    queryCollection(TypesRecord.collection, TypesRecord.serializer,
-        queryBuilder: queryBuilder, limit: limit, singleRecord: singleRecord);
-
-Stream<List<T>> queryCollection<T>(
-    CollectionReference collection, Serializer<T> serializer,
-    {Query Function(Query) queryBuilder,
-    int limit = -1,
-    bool singleRecord = false}) {
-  final builder = queryBuilder ?? (q) => q;
-  var query = builder(collection);
-  if (limit > 0 || singleRecord) {
-    query = query.limit(singleRecord ? 1 : limit);
-  }
-  return query.snapshots().map((s) => s.docs
-      .map(
-        (d) => safeGet(
-          () => serializers.deserializeWith(serializer, serializedData(d)),
-          (e) => print('Error serializing doc ${d.reference.path}:\n$e'),
-        ),
-      )
-      .where((d) => d != null)
-      .toList());
-}
-
-// Creates a Firestore record representing the logged in user if it doesn't yet exist
-Future maybeCreateUser(User user) async {
-  final userRecord = UsersRecord.collection.doc(user.uid);
-  final userExists = await userRecord.get().then((u) => u.exists);
-  if (userExists) {
-    return;
-  }
-
-  final userData = createUsersRecordData(
-    email: user.email,
-    displayName: user.displayName,
-    photoUrl: user.photoURL,
-    uid: user.uid,
-    phoneNumber: user.phoneNumber,
-    createdTime: getCurrentTimestamp,
-  );
-
-  await userRecord.set(userData);
+abstract class BackendService {
+  factory BackendService() => BackendService();
+  static set secret(BackendSecret backendSecret) {}
+  static set repository(BackendRepository backendRepository) {}
+  BackendService.initialize();
 }
